+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "expoServerPort": 19000,
   "devToolsPort": 19002,
   "packagerPort": 19004,
-  "packagerPid": 68577,
-  "expoServerNgrokUrl": "https://ah-uue.kronicle114.kitty.exp.direct",
-  "packagerNgrokUrl": "https://packager.ah-uue.kronicle114.kitty.exp.direct",
-  "ngrokPid": 68592
+  "packagerPid": 77532,
+  "expoServerNgrokUrl": "https://f3-smg.kronicle114.kitty.exp.direct",
+  "packagerNgrokUrl": "https://packager.f3-smg.kronicle114.kitty.exp.direct",
+  "ngrokPid": 77547
 }
