+++ .expo/settings.json
@@ -3,5 +3,5 @@
   "lanType": "ip",
   "dev": true,
   "minify": false,
-  "urlRandomness": "ah-uue"
+  "urlRandomness": "f3-smg"
 }
